refactor(mobile): split personalization settings
